+++ buildSrc/src/main/kotlin/ProjectProperties.kt
@@ -9,7 +9,7 @@ object ProjectProperties {
     const val COMPILE_SDK = 33
     const val MIN_SDK = 24
 
-    const val VERSION_CODE = 4
+    const val VERSION_CODE = 5
     const val VERSION_NAME = "1.0.0"
 
     const val TEST_RUNNER = "androidx.test.runner.AndroidJUnitRunner"
